feat(forms): merged both submit cases into one
