+++ index.html
@@ -4,7 +4,7 @@
 		<meta charset="UTF-8" />
 		<link rel="shortcut icon" href="" type="image/x-icon" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<title>Vite + React + TS</title>
+		<title>Freelegal - Justice, Made Accessible</title>
 	</head>
 	<body>
 		<div id="root"></div>
